Fixed a hover problem with the sub.
Fixed a hover problem with the sub menu of the “nav” component. Created a variable that contains the maximum value for the “z-index” property (the value is decided by the me, the author, it is not the maximum computable). The design of the sub menu and the parent item changed a bit.